+++ src/apis/index.ts
@@ -2,11 +2,13 @@
 /* eslint-disable */
 export * from "./APIKeysApi";
 export * from "./CatalogApi";
+export * from "./ComponentSessionsApi";
+export * from "./HostedSessionsApi";
 export * from "./MembershipsApi";
 export * from "./PlatformOrdersApi";
 export * from "./PlatformWebhooksApi";
 export * from "./PlatformsApi";
-export * from "./PortalSessionsApi";
 export * from "./PracticesApi";
+export * from "./ProviderMappingsApi";
 export * from "./RolesApi";
 export * from "./UsersApi";